+++ examples/WorldClock/Presenter.h
@@ -99,9 +99,8 @@ class Presenter {
       const ZonedDateTime dateTime = ZonedDateTime::forEpochSeconds(
           mRenderingInfo.now, mRenderingInfo.timeZone);
       if (dateTime.isError()) {
-        mOled.println(F("9999-99-99"));
-        mOled.println(F("99:99:99"));
-        mOled.println(F("Error"));
+        clearDisplay();
+        mOled.println(F("<Error>"));
         return;
       }
 
@@ -145,7 +144,7 @@ class Presenter {
       // place name
       mOled.println();
       acetime_t epochSeconds = dateTime.toEpochSeconds();
-      dateTime.timeZone().printAbbrevTo(mOled, epochSeconds);
+      mOled.print(dateTime.timeZone().getAbbrev(epochSeconds));
       mOled.print(' ');
       mOled.print('(');
       mOled.print(mRenderingInfo.name);
@@ -222,7 +221,7 @@ class Presenter {
 
       // abbreviation and place name
       mOled.println();
-      dateTime.timeZone().printAbbrevTo(mOled, mRenderingInfo.now);
+      mOled.print(dateTime.timeZone().getAbbrev(mRenderingInfo.now));
       mOled.print(' ');
       mOled.print('(');
       mOled.print(mRenderingInfo.name);